requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
rembg==2.0.50
segment-anything==1.0
transformers==4.35.2
diffusers==0.24.0
accelerate==0.25.0
xformers==0.0.22
matplotlib==3.7.1
scikit-image==0.21.0
pyyaml==6.0.1
tqdm==4.66.1
huggingface-hub==0.19.4